Fully updated with Auth Services
