+++ .vs/3Sports/v17/DocumentLayout.json
@@ -1,7 +1,16 @@
 {
   "Version": 1,
-  "WorkspaceRootPath": "C:\\repos\\Programs - Practices\\C# Course - 2025\\3Sports\\",
-  "Documents": [],
+  "WorkspaceRootPath": "C:\\Users\\Adich\\Desktop\\3Sports\\",
+  "Documents": [
+    {
+      "AbsoluteMoniker": "D:0:0:{ED46DD07-55F4-4591-A102-F31E0377E7E6}|3Sports.csproj|c:\\users\\adich\\desktop\\3sports\\docs\\database guide.md||{EFC0BB08-EA7D-40C6-A696-C870411A895B}",
+      "RelativeMoniker": "D:0:0:{ED46DD07-55F4-4591-A102-F31E0377E7E6}|3Sports.csproj|solutionrelative:docs\\database guide.md||{EFC0BB08-EA7D-40C6-A696-C870411A895B}"
+    },
+    {
+      "AbsoluteMoniker": "D:0:0:{ED46DD07-55F4-4591-A102-F31E0377E7E6}|3Sports.csproj|c:\\users\\adich\\desktop\\3sports\\services\\teamperfanalysisservice.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
+      "RelativeMoniker": "D:0:0:{ED46DD07-55F4-4591-A102-F31E0377E7E6}|3Sports.csproj|solutionrelative:services\\teamperfanalysisservice.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}"
+    }
+  ],
   "DocumentGroupContainers": [
     {
       "Orientation": 0,
@@ -9,7 +18,7 @@
       "DocumentGroups": [
         {
           "DockedWidth": 200,
-          "SelectedChildIndex": -1,
+          "SelectedChildIndex": 3,
           "Children": [
             {
               "$type": "Bookmark",
@@ -18,6 +27,32 @@
             {
               "$type": "Bookmark",
               "Name": "ST:0:0:{1c4feeaa-4718-4aa9-859d-94ce25d182ba}"
+            },
+            {
+              "$type": "Document",
+              "DocumentIndex": 1,
+              "Title": "TeamPerfAnalysisService.cs",
+              "DocumentMoniker": "C:\\Users\\Adich\\Desktop\\3Sports\\Services\\TeamPerfAnalysisService.cs",
+              "RelativeDocumentMoniker": "Services\\TeamPerfAnalysisService.cs",
+              "ToolTip": "C:\\Users\\Adich\\Desktop\\3Sports\\Services\\TeamPerfAnalysisService.cs",
+              "RelativeToolTip": "Services\\TeamPerfAnalysisService.cs",
+              "ViewState": "AgIAAEkAAAAAAAAAAAAuwFsAAAAAAAAAAAAAAA==",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000738|",
+              "WhenOpened": "2025-03-11T15:43:43.618Z",
+              "EditorCaption": ""
+            },
+            {
+              "$type": "Document",
+              "DocumentIndex": 0,
+              "Title": "Database Guide.md",
+              "DocumentMoniker": "C:\\Users\\Adich\\Desktop\\3Sports\\docs\\Database Guide.md",
+              "RelativeDocumentMoniker": "docs\\Database Guide.md",
+              "ToolTip": "C:\\Users\\Adich\\Desktop\\3Sports\\docs\\Database Guide.md",
+              "RelativeToolTip": "docs\\Database Guide.md",
+              "ViewState": "AgIAAAAAAAAAAAAAAAAAABEAAAAAAAAAAAAAAA==",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.001818|",
+              "WhenOpened": "2025-03-11T15:41:39.789Z",
+              "EditorCaption": ""
             }
           ]
         }
