--- Services/Info.txt
@@ -1,1 +1,0 @@
-﻿> Holds up the business logic like Auth, Filtering, etc.
