+++ classes/User.cs
@@ -9,5 +9,6 @@ namespace _3Sports.classes
         public string Email { get; set; }
         public string PasswordHash { get; set; }
         public DateTime CreatedAt { get; set; } = DateTime.Now;
+        public DateTime? LastLoggedIn { get; set; }
     }
 }
