--- data/Info.txt
@@ -1,2 +1,0 @@
-﻿ > Data Folder will hold the database connections
- and the data files if created.
